distributed resources.py over several files
