+++ modeltestSDK/resources/__init__.py
@@ -1,0 +1,8 @@
+from .campaign import Campaign, Campaigns
+from .floater_config import FloaterConfig, FloaterConfigs
+from .sensor import Sensor, Sensors
+from .test import Test, FloaterTest, WindCalibration, WaveCalibration, Tests
+from .timeseries import TimeSeries, TimeSeriesList
+from .tag import Tag, Tags
+from .statistics import Statistics
+
