+++ modeltestSDK/resources/floater_config.py
@@ -1,0 +1,15 @@
+from typing import Optional
+from .base import Resource, Resources
+
+
+class FloaterConfig(Resource):
+    id: Optional[str]
+    name: str
+    description: str
+    campaign_id: str
+    characteristic_length: float
+    draft: float
+
+
+class FloaterConfigs(Resources[FloaterConfig]):
+    pass
